+++ src/commonMain/kotlin/org/kotools/types/web/DomainName.kt
@@ -1,0 +1,13 @@
+package org.kotools.types.web
+
+import kotools.types.experimental.ExperimentalKotoolsTypesApi
+import kotools.types.internal.ExperimentalSince
+import kotools.types.internal.KotoolsTypesVersion
+
+/**
+ * Represents a
+ * [domain name in the Internet](https://en.wikipedia.org/wiki/Domain_name).
+ */
+@ExperimentalKotoolsTypesApi
+@ExperimentalSince(KotoolsTypesVersion.Unreleased)
+public class DomainName private constructor()